+++ src/components/layout.js
@@ -5,7 +5,7 @@
  * See: https://www.gatsbyjs.com/docs/use-static-query/
  */
 
-import * as React from "react"
+import React, { useEffect } from "react"
 import PropTypes from "prop-types"
 import { useStaticQuery, graphql } from "gatsby"
 
@@ -14,6 +14,12 @@ import Footer from "./footer"
 import "./layout.css"
 
 const Layout = ({ children }) => {
+  // Initialize Google Optimize on 'optimize.activate' event
+  useEffect(() => {
+    window.dataLayer = window.dataLayer || []
+    window.dataLayer.push({ event: "optimize.activate" })
+  }, [])
+
   const data = useStaticQuery(graphql`
     query SiteTitleQuery {
       site {