+++ CHANGELOG.md
@@ -12,6 +12,7 @@ modified Semantic Versioning scheme. See the "Versioning scheme" section of the
 ## [Unreleased]
 
 ### Added
+- BCBSALLY-2446: Add Google Optimize Script for AB Testing
 
 ### Changed